img problem

+++ src/actions/auth.action.js
@@ -90,11 +90,28 @@ export const isUserLoggedIn = () => {
   };
 };
 
+// export const signout = () => {
+//   return async (dispatch) => {
+//     dispatch({ type: authConstants.LOGOUT_REQUEST });
+//     localStorage.clear();
+//     dispatch({ type: authConstants.LOGOUT_SUCCESS });
+//     dispatch({ type: cartConstants.RESET_CART });
+//   };
+// };
+
 export const signout = () => {
   return async (dispatch) => {
     dispatch({ type: authConstants.LOGOUT_REQUEST });
-    localStorage.clear();
-    dispatch({ type: authConstants.LOGOUT_SUCCESS });
-    dispatch({ type: cartConstants.RESET_CART });
+    const res = await axios.post(`/signout`);
+
+    if (res.status === 200) {
+      localStorage.clear();
+      dispatch({ type: authConstants.LOGOUT_SUCCESS });
+    } else {
+      dispatch({
+        type: authConstants.LOGOUT_FAILURE,
+        payload: { error: res.data.error },
+      });
+    }
   };
 };
